requirements.txt
playwright>=1.40.0
playwright-stealth>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=4.0.0
google-genai>=1.0.0
groq>=0.4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
